refactor: extract bloom filter logic from reader/reader.go
Move bloom filter check and detection into reader_bloom.go; split matching
tests into reader_bloom_test.go.
